+++ ebcc/cc/__init__.py
@@ -1,0 +1,1 @@
+"""Coupled cluster solvers."""
